refactor v1.0 finish
