+++ src/configs/config.py
@@ -1,59 +1,52 @@
 import itertools as it
 
-class BaseConfig(object):
-    def __init__(self) -> None:
-        self.get_configs_dict()
-
-    def get_configs_dict(self):
-        configs_dict = {}
-        configs = vars(self)
-        for config_name, config_value in configs.items():
-            if config_name.startswith('__') and config_name.endswith('__'):
-                continue
-            configs_dict[config_name] = config_value
-        self.configs_dict = configs_dict
+def get_configs_dict(configs):
+    configs_dict = {}
+    for config_name, config_value in configs.items():
+        if config_name.startswith('__') and config_name.endswith('__'):
+            continue
+        configs_dict[config_name] = config_value
+    return configs_dict
 
 # ======================================================== #
 #                 Experiment Group Configs                 #
 # ======================================================== #
-class ExptGroupConfig1(BaseConfig):
+class ExptGroupConfig1(object):
     group_name = ["main"]
     dataset = ["CIFAR10"]
     net = ["LeNet5"]
     iteration = [100]
     algo = [{"name": "FedAvg", "param": {"K": 5}}]
     log_freq = [5]
+    # ---------------------------------------------------- #
+    configs_dict = get_configs_dict(locals())
 
-    def __init__(self) -> None:
-        super().__init__()
-
-class ExptGroupConfig2(BaseConfig):
+class ExptGroupConfig2(object):
     group_name = ["main"]
     dataset = ["CIFAR10"]
     net = ["LeNet5"]
     iteration = [100]
     algo = [{"name": "none", "param": {}}]
     log_freq = [5]
-
-    def __init__(self) -> None:
-        super().__init__()
+    # ---------------------------------------------------- #
+    configs_dict = get_configs_dict(locals())
 
 class ExptGroupConfigManager(object):
-    expt_group = [
-        ExptGroupConfig1().get_configs_dict
+    expt_groups = [
+        ExptGroupConfig1.configs_dict
     ]
     def get_expt_groups_configs(self):
         expt_groups_configs = {}
         for expt_group in self.expt_groups:
             combinations = it.product(*(expt_group[name] for name in expt_group))
             expt_group_configs = [{key : value[i] for i,key in enumerate(expt_group)}for value in combinations]
-            expt_groups_configs[expt_group['name'][0]] = expt_group_configs
+            expt_groups_configs[expt_group['group_name'][0]] = expt_group_configs
         return expt_groups_configs
 
 # ======================================================== #
 #                       Global Config                      #
 # ======================================================== #
-class GlobalConfig(BaseConfig):
+class GlobalConfig(object):
     # -------------------- environment ------------------- #
     expt_name = None
     data_path = 'data/'
@@ -74,13 +67,12 @@ class GlobalConfig(BaseConfig):
         "cus_distribution": [5,5,5]
     }
     # ---------------------------------------------------- #
-    def __init__(self) -> None:
-        super().__init__()
+    configs_dict = get_configs_dict(locals())
 
 # ======================================================== #
 #                       Model Config                       #
 # ======================================================== #
-class ModelConfig(BaseConfig):
+class ModelConfig(object):
     optimizer = "Adam"
     scheduler = {
         "name": "StepLR",
@@ -93,8 +85,7 @@ class ModelConfig(BaseConfig):
     min_lr = 0.0001
     batchsize = 64
     # ---------------------------------------------------- #
-    def __init__(self) -> None:
-        super().__init__()
+    configs_dict = get_configs_dict(locals())
 
 expt_group_config_manager = ExptGroupConfigManager()
 global_config = GlobalConfig()
