+++ src/nodes/runner/__init__.py
@@ -2,7 +2,7 @@ from .server import *
 from .client import *
 
 def get_server_runner(hp):
-    algo = hp["algo"]
+    algo = hp["algo"]["name"]
     if algo == "None":
         run_func = base_run_server
     elif algo == "FedAvg":
@@ -13,7 +13,7 @@ def get_server_runner(hp):
     return run_func
 
 def get_client_runner(hp):
-    algo = hp["algo"]
+    algo = hp["algo"]["name"]
     if algo == "None":
         run_func = base_run_client
     elif algo == "FedAvg":
